Clean up format

+++ SendMoney.sol
@@ -1,23 +1,22 @@
 pragma solidity ^0.5.13;
 
 contract SendMoneyExample {
-    
-    uint public balanceReceived;
-    
+    uint256 public balanceReceived;
+
     function receiveMoney() public payable {
         balanceReceived += msg.value;
     }
-    
-    function getBalance() public view returns(uint) {
+
+    function getBalance() public view returns (uint256) {
         return address(this).balance;
     }
-    
+
     function withdrawMoney() public {
         address payable to = msg.sender;
-        
+
         to.transfer(this.getBalance());
     }
-    
+
     function withdrawMoneyTo(address payable _to) public {
         _to.transfer(this.getBalance());
     }
